feat: config static site generation for post detail page
