+++ app/page.tsx
@@ -1,36 +1,9 @@
-import fs from 'fs';
-import matter from 'gray-matter';
-import Link from 'next/link';
-
-import { PostMetaData } from '../models/PostMetaData';
-
-const getPostMetaData = (): PostMetaData[] => {
-    const folder = 'posts/';
-    const files = fs.readdirSync(folder);
-    const markdownPosts = files.filter((file) => file.endsWith('.md'));
-
-    const posts = markdownPosts.map((fileName) => {
-        const fileContent = fs.readFileSync(`posts/${fileName}`, 'utf8');
-        const matterResult = matter(fileContent);
-
-        return {
-            title: matterResult.data.title,
-            difficult: matterResult.data.difficult,
-            slug: fileName.replace('.md', ''),
-        };
-    });
-
-    return posts;
-};
+import PostPreview from '@/components/PostPreview';
+import { getPostMetaData } from '@/lib/utils';
 
 const Homepage = () => {
     const postMetaData = getPostMetaData();
-    const postPreviews = postMetaData.map((post, index) => (
-        <div key={index}>
-            <Link href={`/posts/${post.slug}`}>{post.title}</Link>
-            <p>{post.difficult}</p>
-        </div>
-    ));
+    const postPreviews = postMetaData.map((post) => <PostPreview key={post.slug} {...post} />);
 
     return <div>{postPreviews}</div>;
 };
