+++ posts/7-reverse-integer.md
@@ -1,0 +1,33 @@
+---
+title: "7. Reverse Integer"
+difficult: "Medium"
+---
+
+Given a signed 32-bit integer `x`, return `x` with its digits reversed. If reversing `x` causes the value to go outside the signed 32-bit integer range `[-231, 231 - 1]`, then return `0`.
+
+**Assume the environment does not allow you to store 64-bit integers (signed or unsigned).**
+
+**Example 1:**
+
+```text
+Input: x = 123
+Output: 321
+```
+
+**Example 2:**
+
+```text
+Input: x = -123
+Output: -321
+```
+
+**Example 3:**
+
+```text
+Input: x = 120
+Output: 21
+```
+
+**Constraints:**
+
+- -231 <= x <= 231 - 1
