+++ tsconfig.json
@@ -1,5 +1,12 @@
 {
   "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@/app/*": ["app/*"],
+      "@/components/*": ["components/*"],
+      "@/lib/*": ["lib/*"],
+      "@/models/*": ["models/*"],
+    },
     "target": "es5",
     "lib": [
       "dom",
